+++ Source/SwiftLintFramework/Models/RuleDescription.swift
@@ -3,7 +3,7 @@
 //  SwiftLint
 //
 //  Created by Chris Eidhof on 25/05/15.
-//  Copyright (c) 2015 Realm. All rights reserved.
+//  Copyright © 2015 Realm. All rights reserved.
 //
 
 public struct RuleDescription: Equatable {