+++ Source/SwiftLintFramework/Rules/PrivateOutletRule.swift
@@ -2,7 +2,7 @@
 //  PrivateOutletRule.swift
 //  SwiftLint
 //
-//  Created by Olivier Halligon on 12/08/2016.
+//  Created by Olivier Halligon on 12/8/16.
 //  Copyright © 2016 Realm. All rights reserved.
 //